@@ -1,13 +1,11 @@
 package filesystem
 
 import (
-	"bufio"
+	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
-	"path"
 	"path/filepath"
-	"sort"
+	"slices"
 	"strconv"
 	"strings"
 	"sync"
@@ -15,220 +13,208 @@ import (
 	"time"
 
 	"emperror.dev/errors"
+	"github.com/apex/log"
 	"github.com/gabriel-vasile/mimetype"
-	"github.com/karrick/godirwalk"
 	ignore "github.com/sabhiram/go-gitignore"
 
 	"github.com/Tech-Gamer/nwy-wings/config"
-	"github.com/Tech-Gamer/nwy-wings/system"
+	"github.com/Tech-Gamer/nwy-wings/internal/ufs"
 )
 
 type Filesystem struct {
+	unixFS *ufs.Quota
+
 	mu                sync.RWMutex
 	lastLookupTime    *usageLookupTime
-	lookupInProgress  *system.AtomicBool
-	diskUsed          int64
+	lookupInProgress  atomic.Bool
 	diskCheckInterval time.Duration
 	denylist          *ignore.GitIgnore
-
-	// The maximum amount of disk space (in bytes) that this Filesystem instance can use.
-	diskLimit int64
-
-	// The root data directory path for this Filesystem instance.
-	root string
 
 	isTest bool
 }
 
 // New creates a new Filesystem instance for a given server.
-func New(root string, size int64, denylist []string) *Filesystem {
+func New(root string, size int64, denylist []string) (*Filesystem, error) {
+	if err := os.MkdirAll(root, 0o755); err != nil {
+		return nil, err
+	}
+	unixFS, err := ufs.NewUnixFS(root, config.UseOpenat2())
+	if err != nil {
+		return nil, err
+	}
+	quota := ufs.NewQuota(unixFS, size)
+
 	return &Filesystem{
-		root:              root,
-		diskLimit:         size,
+		unixFS: quota,
+
 		diskCheckInterval: time.Duration(config.Get().System.DiskCheckInterval),
 		lastLookupTime:    &usageLookupTime{},
-		lookupInProgress:  system.NewAtomicBool(false),
 		denylist:          ignore.CompileIgnoreLines(denylist...),
-	}
+	}, nil
 }
 
 // Path returns the root path for the Filesystem instance.
 func (fs *Filesystem) Path() string {
-	return fs.root
+	return fs.unixFS.BasePath()
+}
+
+// ReadDir reads directory entries.
+func (fs *Filesystem) ReadDir(path string) ([]ufs.DirEntry, error) {
+	return fs.unixFS.ReadDir(path)
+}
+
+// ReadDirStat is like ReadDir except that it returns FileInfo for each entry
+// instead of just a DirEntry.
+func (fs *Filesystem) ReadDirStat(path string) ([]ufs.FileInfo, error) {
+	return ufs.ReadDirMap(fs.unixFS.UnixFS, path, func(e ufs.DirEntry) (ufs.FileInfo, error) {
+		return e.Info()
+	})
 }
 
 // File returns a reader for a file instance as well as the stat information.
-func (fs *Filesystem) File(p string) (*os.File, Stat, error) {
-	cleaned, err := fs.SafePath(p)
+func (fs *Filesystem) File(p string) (ufs.File, Stat, error) {
+	f, err := fs.unixFS.Open(p)
 	if err != nil {
-		return nil, Stat{}, errors.WithStackIf(err)
+		return nil, Stat{}, err
 	}
-	st, err := fs.Stat(cleaned)
+	st, err := statFromFile(f)
 	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return nil, Stat{}, newFilesystemError(ErrNotExist, err)
-		}
-		return nil, Stat{}, errors.WithStackIf(err)
-	}
-	if st.IsDir() {
-		return nil, Stat{}, newFilesystemError(ErrCodeIsDirectory, nil)
-	}
-	f, err := os.Open(cleaned)
-	if err != nil {
-		return nil, Stat{}, errors.WithStackIf(err)
+		_ = f.Close()
+		return nil, Stat{}, err
 	}
 	return f, st, nil
+}
+
+func (fs *Filesystem) UnixFS() *ufs.UnixFS {
+	return fs.unixFS.UnixFS
 }
 
 // Touch acts by creating the given file and path on the disk if it is not present
 // already. If  it is present, the file is opened using the defaults which will truncate
 // the contents. The opened file is then returned to the caller.
-func (fs *Filesystem) Touch(p string, flag int) (*os.File, error) {
-	cleaned, err := fs.SafePath(p)
-	if err != nil {
-		return nil, err
-	}
-	f, err := os.OpenFile(cleaned, flag, 0o644)
-	if err == nil {
-		return f, nil
-	}
-	if f != nil {
-		_ = f.Close()
-	}
-	// If the error is not because it doesn't exist then we just need to bail at this point.
-	if !errors.Is(err, os.ErrNotExist) {
-		return nil, errors.Wrap(err, "server/filesystem: touch: failed to open file handle")
-	}
-	// Only create and chown the directory if it doesn't exist.
-	if _, err := os.Stat(filepath.Dir(cleaned)); errors.Is(err, os.ErrNotExist) {
-		// Create the path leading up to the file we're trying to create, setting the final perms
-		// on it as we go.
-		if err := os.MkdirAll(filepath.Dir(cleaned), 0o755); err != nil {
-			return nil, errors.Wrap(err, "server/filesystem: touch: failed to create directory tree")
-		}
-		if err := fs.Chown(filepath.Dir(cleaned)); err != nil {
-			return nil, err
-		}
-	}
-	o := &fileOpener{}
-	// Try to open the file now that we have created the pathing necessary for it, and then
-	// Chown that file so that the permissions don't mess with things.
-	f, err = o.open(cleaned, flag, 0o644)
-	if err != nil {
-		return nil, errors.Wrap(err, "server/filesystem: touch: failed to open file with wait")
-	}
-	_ = fs.Chown(cleaned)
-	return f, nil
+func (fs *Filesystem) Touch(p string, flag int) (ufs.File, error) {
+	return fs.unixFS.Touch(p, flag, 0o644)
 }
 
 // Writefile writes a file to the system. If the file does not already exist one
 // will be created. This will also properly recalculate the disk space used by
 // the server when writing new files or modifying existing ones.
+//
+// DEPRECATED: use `Write` instead.
 func (fs *Filesystem) Writefile(p string, r io.Reader) error {
-	cleaned, err := fs.SafePath(p)
-	if err != nil {
-		return err
-	}
-
 	var currentSize int64
-	// If the file does not exist on the system already go ahead and create the pathway
-	// to it and an empty file. We'll then write to it later on after this completes.
-	stat, err := os.Stat(cleaned)
-	if err != nil && !os.IsNotExist(err) {
+	st, err := fs.unixFS.Stat(p)
+	if err != nil && !errors.Is(err, ufs.ErrNotExist) {
 		return errors.Wrap(err, "server/filesystem: writefile: failed to stat file")
 	} else if err == nil {
-		if stat.IsDir() {
-			return errors.WithStack(&Error{code: ErrCodeIsDirectory, resolved: cleaned})
+		if st.IsDir() {
+			// TODO: resolved
+			return errors.WithStack(&Error{code: ErrCodeIsDirectory, resolved: ""})
 		}
-		currentSize = stat.Size()
+		currentSize = st.Size()
 	}
 
-	br := bufio.NewReader(r)
+	// Touch the file and return the handle to it at this point. This will
+	// create or truncate the file, and create any necessary parent directories
+	// if they are missing.
+	file, err := fs.unixFS.Touch(p, ufs.O_RDWR|ufs.O_TRUNC, 0o644)
+	if err != nil {
+		return fmt.Errorf("error touching file: %w", err)
+	}
+	defer file.Close()
+
+	// Do not use CopyBuffer here, it is wasteful as the file implements
+	// io.ReaderFrom, which causes it to not use the buffer anyways.
+	n, err := io.Copy(file, r)
+
+	// Adjust the disk usage to account for the old size and the new size of the file.
+	fs.unixFS.Add(n - currentSize)
+
+	if err := fs.chownFile(p); err != nil {
+		return fmt.Errorf("error chowning file: %w", err)
+	}
+	// Return the error from io.Copy.
+	return err
+}
+
+func (fs *Filesystem) Write(p string, r io.Reader, newSize int64, mode ufs.FileMode) error {
+	var currentSize int64
+	st, err := fs.unixFS.Stat(p)
+	if err != nil && !errors.Is(err, ufs.ErrNotExist) {
+		return errors.Wrap(err, "server/filesystem: writefile: failed to stat file")
+	} else if err == nil {
+		if st.IsDir() {
+			// TODO: resolved
+			return errors.WithStack(&Error{code: ErrCodeIsDirectory, resolved: ""})
+		}
+		currentSize = st.Size()
+	}
+
 	// Check that the new size we're writing to the disk can fit. If there is currently
 	// a file we'll subtract that current file size from the size of the buffer to determine
 	// the amount of new data we're writing (or amount we're removing if smaller).
-	if err := fs.HasSpaceFor(int64(br.Size()) - currentSize); err != nil {
+	if err := fs.HasSpaceFor(newSize - currentSize); err != nil {
 		return err
 	}
 
-	// Touch the file and return the handle to it at this point. This will create the file,
-	// any necessary directories, and set the proper owner of the file.
-	file, err := fs.Touch(cleaned, os.O_RDWR|os.O_CREATE|os.O_TRUNC)
+	// Touch the file and return the handle to it at this point. This will
+	// create or truncate the file, and create any necessary parent directories
+	// if they are missing.
+	file, err := fs.unixFS.Touch(p, ufs.O_RDWR|ufs.O_TRUNC, mode)
 	if err != nil {
 		return err
 	}
 	defer file.Close()
 
-	buf := make([]byte, 1024*4)
-	sz, err := io.CopyBuffer(file, r, buf)
+	if newSize == 0 {
+		// Subtract the previous size of the file if the new size is 0.
+		fs.unixFS.Add(-currentSize)
+	} else {
+		// Do not use CopyBuffer here, it is wasteful as the file implements
+		// io.ReaderFrom, which causes it to not use the buffer anyways.
+		var n int64
+		n, err = io.Copy(file, io.LimitReader(r, newSize))
 
-	// Adjust the disk usage to account for the old size and the new size of the file.
-	fs.addDisk(sz - currentSize)
+		// Adjust the disk usage to account for the old size and the new size of the file.
+		fs.unixFS.Add(n - currentSize)
+	}
 
-	return fs.unsafeChown(cleaned)
-}
-
-// Creates a new directory (name) at a specified path (p) for the server.
-func (fs *Filesystem) CreateDirectory(name string, p string) error {
-	cleaned, err := fs.SafePath(path.Join(p, name))
-	if err != nil {
+	if err := fs.chownFile(p); err != nil {
 		return err
 	}
-	return os.MkdirAll(cleaned, 0o755)
+	// Return any remaining error.
+	return err
 }
 
-// Rename moves (or renames) a file or directory.
-func (fs *Filesystem) Rename(from string, to string) error {
-	cleanedFrom, err := fs.SafePath(from)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-
-	cleanedTo, err := fs.SafePath(to)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-
-	// If the target file or directory already exists the rename function will fail, so just
-	// bail out now.
-	if _, err := os.Stat(cleanedTo); err == nil {
-		return os.ErrExist
-	}
-
-	if cleanedTo == fs.Path() {
-		return errors.New("attempting to rename into an invalid directory space")
-	}
-
-	d := strings.TrimSuffix(cleanedTo, path.Base(cleanedTo))
-	// Ensure that the directory we're moving into exists correctly on the system. Only do this if
-	// we're not at the root directory level.
-	if d != fs.Path() {
-		if mkerr := os.MkdirAll(d, 0o755); mkerr != nil {
-			return errors.WithMessage(mkerr, "failed to create directory structure for file rename")
-		}
-	}
-
-	if err := os.Rename(cleanedFrom, cleanedTo); err != nil {
-		return errors.WithStack(err)
-	}
-	return nil
+// CreateDirectory creates a new directory (name) at a specified path (p) for
+// the server.
+func (fs *Filesystem) CreateDirectory(name string, p string) error {
+	return fs.unixFS.MkdirAll(filepath.Join(p, name), 0o755)
 }
 
-// Recursively iterates over a file or directory and sets the permissions on all of the
+func (fs *Filesystem) Rename(oldpath, newpath string) error {
+	return fs.unixFS.Rename(oldpath, newpath)
+}
+
+func (fs *Filesystem) Symlink(oldpath, newpath string) error {
+	return fs.unixFS.Symlink(oldpath, newpath)
+}
+
+func (fs *Filesystem) chownFile(name string) error {
+	if fs.isTest {
+		return nil
+	}
+
+	uid := config.Get().System.User.Uid
+	gid := config.Get().System.User.Gid
+	return fs.unixFS.Lchown(name, uid, gid)
+}
+
+// Chown recursively iterates over a file or directory and sets the permissions on all of the
 // underlying files. Iterate over all of the files and directories. If it is a file just
 // go ahead and perform the chown operation. Otherwise dig deeper into the directory until
 // we've run out of directories to dig into.
-func (fs *Filesystem) Chown(path string) error {
-	cleaned, err := fs.SafePath(path)
-	if err != nil {
-		return err
-	}
-	return fs.unsafeChown(cleaned)
-}
-
-// unsafeChown chowns the given path, without checking if the path is safe. This should only be used
-// when the path has already been checked.
-func (fs *Filesystem) unsafeChown(path string) error {
+func (fs *Filesystem) Chown(p string) error {
 	if fs.isTest {
 		return nil
 	}
@@ -236,54 +222,44 @@ func (fs *Filesystem) unsafeChown(path string) error {
 	uid := config.Get().System.User.Uid
 	gid := config.Get().System.User.Gid
 
+	dirfd, name, closeFd, err := fs.unixFS.SafePath(p)
+	defer closeFd()
+	if err != nil {
+		return err
+	}
+
 	// Start by just chowning the initial path that we received.
-	if err := os.Chown(path, uid, gid); err != nil {
+	if err := fs.unixFS.Lchownat(dirfd, name, uid, gid); err != nil {
 		return errors.Wrap(err, "server/filesystem: chown: failed to chown path")
 	}
 
 	// If this is not a directory we can now return from the function, there is nothing
 	// left that we need to do.
-	if st, err := os.Stat(path); err != nil || !st.IsDir() {
+	if st, err := fs.unixFS.Lstatat(dirfd, name); err != nil || !st.IsDir() {
 		return nil
 	}
 
-	// If this was a directory, begin walking over its contents recursively and ensure that all
-	// of the subfiles and directories get their permissions updated as well.
-	err := godirwalk.Walk(path, &godirwalk.Options{
-		Unsorted: true,
-		Callback: func(p string, e *godirwalk.Dirent) error {
-			// Do not attempt to chown a symlink. Go's os.Chown function will affect the symlink
-			// so if it points to a location outside the data directory the user would be able to
-			// (un)intentionally modify that files permissions.
-			if e.IsSymlink() {
-				if e.IsDir() {
-					return godirwalk.SkipThis
-				}
-
-				return nil
-			}
-
-			return os.Chown(p, uid, gid)
-		},
-	})
-	return errors.Wrap(err, "server/filesystem: chown: failed to chown during walk function")
+	// This walker is probably some of the most efficient code in Wings. It has
+	// an internally re-used buffer for listing directory entries and doesn't
+	// need to check if every individual path it touches is safe as the code
+	// doesn't traverse symlinks, is immune to symlink timing attacks, and
+	// gives us a dirfd and file name to make a direct syscall with.
+	if err := fs.unixFS.WalkDirat(dirfd, name, func(dirfd int, name, _ string, info ufs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if err := fs.unixFS.Lchownat(dirfd, name, uid, gid); err != nil {
+			return err
+		}
+		return nil
+	}); err != nil {
+		return fmt.Errorf("server/filesystem: chown: failed to chown during walk function: %w", err)
+	}
+	return nil
 }
 
-func (fs *Filesystem) Chmod(path string, mode os.FileMode) error {
-	cleaned, err := fs.SafePath(path)
-	if err != nil {
-		return err
-	}
-
-	if fs.isTest {
-		return nil
-	}
-
-	if err := os.Chmod(cleaned, mode); err != nil {
-		return err
-	}
-
-	return nil
+func (fs *Filesystem) Chmod(path string, mode ufs.FileMode) error {
+	return fs.unixFS.Chmod(path, mode)
 }
 
 // Begin looping up to 50 times to try and create a unique copy file name. This will take
@@ -294,7 +270,7 @@ func (fs *Filesystem) Chmod(path string, mode os.FileMode) error {
 // Could probably make this more efficient by checking if there are any files matching the copy
 // pattern, and trying to find the highest number and then incrementing it by one rather than
 // looping endlessly.
-func (fs *Filesystem) findCopySuffix(dir string, name string, extension string) (string, error) {
+func (fs *Filesystem) findCopySuffix(dirfd int, name, extension string) (string, error) {
 	var i int
 	suffix := " copy"
 
@@ -306,11 +282,10 @@ func (fs *Filesystem) findCopySuffix(dir string, name string, extension string) 
 		n := name + suffix + extension
 		// If we stat the file and it does not exist that means we're good to create the copy. If it
 		// does exist, we'll just continue to the next loop and try again.
-		if _, err := fs.Stat(path.Join(dir, n)); err != nil {
-			if !errors.Is(err, os.ErrNotExist) {
+		if _, err := fs.unixFS.Lstatat(dirfd, n); err != nil {
+			if !errors.Is(err, ufs.ErrNotExist) {
 				return "", err
 			}
-
 			break
 		}
 
@@ -322,53 +297,68 @@ func (fs *Filesystem) findCopySuffix(dir string, name string, extension string) 
 	return name + suffix + extension, nil
 }
 
-// Copies a given file to the same location and appends a suffix to the file to indicate that
-// it has been copied.
+// Copy copies a given file to the same location and appends a suffix to the
+// file to indicate that it has been copied.
 func (fs *Filesystem) Copy(p string) error {
-	cleaned, err := fs.SafePath(p)
+	dirfd, name, closeFd, err := fs.unixFS.SafePath(p)
+	defer closeFd()
 	if err != nil {
 		return err
 	}
-
-	s, err := os.Stat(cleaned)
-	if err != nil {
-		return err
-	} else if s.IsDir() || !s.Mode().IsRegular() {
-		// If this is a directory or not a regular file, just throw a not-exist error
-		// since anything calling this function should understand what that means.
-		return os.ErrNotExist
-	}
-
-	// Check that copying this file wouldn't put the server over its limit.
-	if err := fs.HasSpaceFor(s.Size()); err != nil {
-		return err
-	}
-
-	base := filepath.Base(cleaned)
-	relative := strings.TrimSuffix(strings.TrimPrefix(cleaned, fs.Path()), base)
-	extension := filepath.Ext(base)
-	name := strings.TrimSuffix(base, extension)
-
-	// Ensure that ".tar" is also counted as apart of the file extension.
-	// There might be a better way to handle this for other double file extensions,
-	// but this is a good workaround for now.
-	if strings.HasSuffix(name, ".tar") {
-		extension = ".tar" + extension
-		name = strings.TrimSuffix(name, ".tar")
-	}
-
-	source, err := os.Open(cleaned)
+	source, err := fs.unixFS.OpenFileat(dirfd, name, ufs.O_RDONLY, 0)
 	if err != nil {
 		return err
 	}
 	defer source.Close()
+	info, err := source.Stat()
+	if err != nil {
+		return err
+	}
+	if info.IsDir() || !info.Mode().IsRegular() {
+		// If this is a directory or not a regular file, just throw a not-exist error
+		// since anything calling this function should understand what that means.
+		return ufs.ErrNotExist
+	}
+	currentSize := info.Size()
 
-	n, err := fs.findCopySuffix(relative, name, extension)
+	// Check that copying this file wouldn't put the server over its limit.
+	if err := fs.HasSpaceFor(currentSize); err != nil {
+		return err
+	}
+
+	base := info.Name()
+	extension := filepath.Ext(base)
+	baseName := strings.TrimSuffix(base, extension)
+
+	// Ensure that ".tar" is also counted as apart of the file extension.
+	// There might be a better way to handle this for other double file extensions,
+	// but this is a good workaround for now.
+	if strings.HasSuffix(baseName, ".tar") {
+		extension = ".tar" + extension
+		baseName = strings.TrimSuffix(baseName, ".tar")
+	}
+
+	newName, err := fs.findCopySuffix(dirfd, baseName, extension)
+	if err != nil {
+		return err
+	}
+	dst, err := fs.unixFS.OpenFileat(dirfd, newName, ufs.O_WRONLY|ufs.O_CREATE, info.Mode())
 	if err != nil {
 		return err
 	}
 
-	return fs.Writefile(path.Join(relative, n), source)
+	// Do not use CopyBuffer here, it is wasteful as the file implements
+	// io.ReaderFrom, which causes it to not use the buffer anyways.
+	n, err := io.Copy(dst, io.LimitReader(source, currentSize))
+	fs.unixFS.Add(n)
+
+	if !fs.isTest {
+		if err := fs.unixFS.Lchownat(dirfd, newName, config.Get().System.User.Uid, config.Get().System.User.Gid); err != nil {
+			return err
+		}
+	}
+	// Return the error from io.Copy.
+	return err
 }
 
 // TruncateRootDirectory removes _all_ files and directories from a server's
@@ -380,211 +370,128 @@ func (fs *Filesystem) TruncateRootDirectory() error {
 	if err := os.Mkdir(fs.Path(), 0o755); err != nil {
 		return err
 	}
-	atomic.StoreInt64(&fs.diskUsed, 0)
+	_ = fs.unixFS.Close()
+	unixFS, err := ufs.NewUnixFS(fs.Path(), config.UseOpenat2())
+	if err != nil {
+		return err
+	}
+	var limit int64
+	if fs.isTest {
+		limit = 0
+	} else {
+		limit = fs.unixFS.Limit()
+	}
+	fs.unixFS = ufs.NewQuota(unixFS, limit)
 	return nil
 }
 
 // Delete removes a file or folder from the system. Prevents the user from
 // accidentally (or maliciously) removing their root server data directory.
 func (fs *Filesystem) Delete(p string) error {
-	// This is one of the few (only?) places in the codebase where we're explicitly not using
-	// the SafePath functionality when working with user provided input. If we did, you would
-	// not be able to delete a file that is a symlink pointing to a location outside the data
-	// directory.
-	//
-	// We also want to avoid resolving a symlink that points _within_ the data directory and thus
-	// deleting the actual source file for the symlink rather than the symlink itself. For these
-	// purposes just resolve the actual file path using filepath.Join() and confirm that the path
-	// exists within the data directory.
-	resolved := fs.unsafeFilePath(p)
-	if !fs.unsafeIsInDataDirectory(resolved) {
-		return NewBadPathResolution(p, resolved)
-	}
-
-	// Block any whoopsies.
-	if resolved == fs.Path() {
-		return errors.New("cannot delete root server directory")
-	}
-
-	st, err := os.Lstat(resolved)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			fs.error(err).Warn("error while attempting to stat file before deletion")
-			return err
-		}
-
-		// The following logic is used to handle a case where a user attempts to
-		// delete a file that does not exist through a directory symlink.
-		// We don't want to reveal that the file does not exist, so we validate
-		// the path of the symlink and return a bad path error if it is invalid.
-
-		// The requested file or directory doesn't exist, so at this point we
-		// need to iterate up the path chain until we hit a directory that
-		// _does_ exist and can be validated.
-		parts := strings.Split(filepath.Dir(resolved), "/")
-
-		// Range over all the path parts and form directory paths from the end
-		// moving up until we have a valid resolution, or we run out of paths to
-		// try.
-		for k := range parts {
-			try := strings.Join(parts[:(len(parts)-k)], "/")
-			if !fs.unsafeIsInDataDirectory(try) {
-				break
-			}
-
-			t, err := filepath.EvalSymlinks(try)
-			if err == nil {
-				if !fs.unsafeIsInDataDirectory(t) {
-					return NewBadPathResolution(p, t)
-				}
-				break
-			}
-		}
-
-		// Always return early if the file does not exist.
-		return nil
-	}
-
-	// If the file is not a symlink, we need to check that it is not within a
-	// symlinked directory that points outside the data directory.
-	if st.Mode()&os.ModeSymlink == 0 {
-		ep, err := filepath.EvalSymlinks(resolved)
-		if err != nil {
-			if !os.IsNotExist(err) {
-				return err
-			}
-		} else if !fs.unsafeIsInDataDirectory(ep) {
-			return NewBadPathResolution(p, ep)
-		}
-	}
-
-	if st.IsDir() {
-		if s, err := fs.DirectorySize(resolved); err == nil {
-			fs.addDisk(-s)
-		}
-	} else {
-		fs.addDisk(-st.Size())
-	}
-
-	return os.RemoveAll(resolved)
+	return fs.unixFS.RemoveAll(p)
 }
 
-type fileOpener struct {
-	busy uint
-}
-
-// Attempts to open a given file up to "attempts" number of times, using a backoff. If the file
-// cannot be opened because of a "text file busy" error, we will attempt until the number of attempts
-// has been exhaused, at which point we will abort with an error.
-func (fo *fileOpener) open(path string, flags int, perm os.FileMode) (*os.File, error) {
-	for {
-		f, err := os.OpenFile(path, flags, perm)
-
-		// If there is an error because the text file is busy, go ahead and sleep for a few
-		// hundred milliseconds and then try again up to three times before just returning the
-		// error back to the caller.
-		//
-		// Based on code from: https://github.com/golang/go/issues/22220#issuecomment-336458122
-		if err != nil && fo.busy < 3 && strings.Contains(err.Error(), "text file busy") {
-			time.Sleep(100 * time.Millisecond << fo.busy)
-			fo.busy++
-			continue
-		}
-
-		return f, err
-	}
-}
+//type fileOpener struct {
+//	fs   *Filesystem
+//	busy uint
+//}
+//
+//// Attempts to open a given file up to "attempts" number of times, using a backoff. If the file
+//// cannot be opened because of a "text file busy" error, we will attempt until the number of attempts
+//// has been exhaused, at which point we will abort with an error.
+//func (fo *fileOpener) open(path string, flags int, perm ufs.FileMode) (ufs.File, error) {
+//	for {
+//		f, err := fo.fs.unixFS.OpenFile(path, flags, perm)
+//
+//		// If there is an error because the text file is busy, go ahead and sleep for a few
+//		// hundred milliseconds and then try again up to three times before just returning the
+//		// error back to the caller.
+//		//
+//		// Based on code from: https://github.com/golang/go/issues/22220#issuecomment-336458122
+//		if err != nil && fo.busy < 3 && strings.Contains(err.Error(), "text file busy") {
+//			time.Sleep(100 * time.Millisecond << fo.busy)
+//			fo.busy++
+//			continue
+//		}
+//
+//		return f, err
+//	}
+//}
 
 // ListDirectory lists the contents of a given directory and returns stat
 // information about each file and folder within it.
 func (fs *Filesystem) ListDirectory(p string) ([]Stat, error) {
-	cleaned, err := fs.SafePath(p)
-	if err != nil {
-		return nil, err
-	}
-
-	files, err := ioutil.ReadDir(cleaned)
-	if err != nil {
-		return nil, err
-	}
-
-	var wg sync.WaitGroup
-
-	// You must initialize the output of this directory as a non-nil value otherwise
-	// when it is marshaled into a JSON object you'll just get 'null' back, which will
-	// break the panel badly.
-	out := make([]Stat, len(files))
-
-	// Iterate over all of the files and directories returned and perform an async process
-	// to get the mime-type for them all.
-	for i, file := range files {
-		wg.Add(1)
-
-		go func(idx int, f os.FileInfo) {
-			defer wg.Done()
-
-			var m *mimetype.MIME
-			d := "inode/directory"
-			if !f.IsDir() {
-				cleanedp := filepath.Join(cleaned, f.Name())
-				if f.Mode()&os.ModeSymlink != 0 {
-					cleanedp, _ = fs.SafePath(filepath.Join(cleaned, f.Name()))
-				}
-
-				// Don't try to detect the type on a pipe — this will just hang the application and
-				// you'll never get a response back.
-				//
-				// @see https://github.com/pterodactyl/panel/issues/4059
-				if cleanedp != "" && f.Mode()&os.ModeNamedPipe == 0 {
-					m, _ = mimetype.DetectFile(filepath.Join(cleaned, f.Name()))
-				} else {
-					// Just pass this for an unknown type because the file could not safely be resolved within
-					// the server data path.
-					d = "application/octet-stream"
-				}
-			}
-
-			st := Stat{FileInfo: f, Mimetype: d}
-			if m != nil {
-				st.Mimetype = m.String()
-			}
-			out[idx] = st
-		}(i, file)
-	}
-
-	wg.Wait()
-
-	// Sort the output alphabetically to begin with since we've run the output
-	// through an asynchronous process and the order is gonna be very random.
-	sort.SliceStable(out, func(i, j int) bool {
-		if out[i].Name() == out[j].Name() || out[i].Name() > out[j].Name() {
-			return true
+	// Read entries from the path on the filesystem, using the mapped reader, so
+	// we can map the DirEntry slice into a Stat slice with mimetype information.
+	out, err := ufs.ReadDirMap(fs.unixFS.UnixFS, p, func(e ufs.DirEntry) (Stat, error) {
+		info, err := e.Info()
+		if err != nil {
+			return Stat{}, err
 		}
-		return false
+
+		var d string
+		if e.Type().IsDir() {
+			d = "inode/directory"
+		} else {
+			d = "application/octet-stream"
+		}
+		var m *mimetype.MIME
+		if e.Type().IsRegular() {
+			// TODO: I should probably find a better way to do this.
+			eO := e.(interface {
+				Open() (ufs.File, error)
+			})
+			f, err := eO.Open()
+			if err != nil {
+				return Stat{}, err
+			}
+			m, err = mimetype.DetectReader(f)
+			if err != nil {
+				log.Error(err.Error())
+			}
+			_ = f.Close()
+		}
+
+		st := Stat{FileInfo: info, Mimetype: d}
+		if m != nil {
+			st.Mimetype = m.String()
+		}
+		return st, nil
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	// Sort entries alphabetically.
+	slices.SortStableFunc(out, func(a, b Stat) int {
+		switch {
+		case a.Name() == b.Name():
+			return 0
+		case a.Name() > b.Name():
+			return 1
+		default:
+			return -1
+		}
 	})
 
-	// Then, sort it so that directories are listed first in the output. Everything
-	// will continue to be alphabetized at this point.
-	sort.SliceStable(out, func(i, j int) bool {
-		return out[i].IsDir()
+	// Sort folders before other file types.
+	slices.SortStableFunc(out, func(a, b Stat) int {
+		switch {
+		case a.IsDir() && b.IsDir():
+			return 0
+		case a.IsDir():
+			return 1
+		default:
+			return -1
+		}
 	})
 
 	return out, nil
 }
 
 func (fs *Filesystem) Chtimes(path string, atime, mtime time.Time) error {
-	cleaned, err := fs.SafePath(path)
-	if err != nil {
-		return err
-	}
-
 	if fs.isTest {
 		return nil
 	}
-
-	if err := os.Chtimes(cleaned, atime, mtime); err != nil {
-		return err
-	}
-
-	return nil
+	return fs.unixFS.Chtimes(path, atime, mtime)
 }
