@@ -2,11 +2,12 @@ package filesystem
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
 
 	"emperror.dev/errors"
 	"github.com/apex/log"
+
+	"github.com/pterodactyl/wings/internal/ufs"
 )
 
 type ErrorCode string
@@ -86,15 +87,15 @@ func (e *Error) Unwrap() error {
 
 // Generates an error logger instance with some basic information.
 func (fs *Filesystem) error(err error) *log.Entry {
-	return log.WithField("subsystem", "filesystem").WithField("root", fs.root).WithField("error", err)
+	return log.WithField("subsystem", "filesystem").WithField("root", fs.Path()).WithField("error", err)
 }
 
 // Handle errors encountered when walking through directories.
 //
 // If there is a path resolution error just skip the item entirely. Only return this for a
 // directory, otherwise return nil. Returning this error for a file will stop the walking
-// for the remainder of the directory. This is assuming an os.FileInfo struct was even returned.
-func (fs *Filesystem) handleWalkerError(err error, f os.FileInfo) error {
+// for the remainder of the directory. This is assuming an FileInfo struct was even returned.
+func (fs *Filesystem) handleWalkerError(err error, f ufs.FileInfo) error {
 	if !IsErrorCode(err, ErrCodePathResolution) {
 		return err
 	}
